Move 'getCustomerPanelSettingsFromWidgetProps' into shared; Add 'customer-panel' class name;
