+++ src/shared/utils/index.js
@@ -4,6 +4,7 @@ export * from './address-utils';
 export * from './element-utils';
 export * from './instance-utils';
 export * from './override-utils';
+export * from './settings-utils';
 export * from './store-utils';
 export * from './validation-utils';
 export * from './widget-utils';
